fix: validate user for put and delete routes with jwt token

+++ controllers/user.controller.js
@@ -39,27 +39,42 @@ export const getParticularUser = async (req, res, next) => {
 }
 
 export const updateUser = async (req, res, next) => {
-    const userExist = await User.findByIdAndUpdate(req.params.id, { ...req.body })
-    if (!userExist)  {
-        const error = new Error("The user doesn't exist")
-        error.statusCode = 404
+    try {
+        const { id } = req.params
+        if (req.user._id != id) {
+            const error = new Error("You can't update someone else' account")
+            error.statusCode = 401
+            throw error
+        }
+        const userExist = await User.findByIdAndUpdate(req.params.id, { ...req.body })
+        if (!userExist)  {
+            const error = new Error("The user doesn't exist")
+            error.statusCode = 404
+            next(error)
+        }
+        
+        const updatedUser = await User.findById(req.params.id)
+        
+        res.status(200).json({
+            success: true,
+            message: 'Your account has been successfully updated',
+            data: {
+                user: updatedUser
+            }
+        })
+    } catch (error) {
         next(error)
     }
-
-    const updatedUser = await User.findById(req.params.id)
-
-    res.status(200).json({
-        success: true,
-        message: 'Your account has been successfully updated',
-        data: {
-            user: updatedUser
-        }
-    })
 }
 
 export const deleteUser = async (req, res, next) => {
     try {
         const { id } = req.params
+        if (req.user._id != id) {
+            const error = new Error("You can't delete someone else' account")
+            error.statusCode = 401
+            throw error
+        }
         // Check if user exist
         const userExist = await User.findByIdAndDelete(id)
